Fix MinGW build warning caused by tzname redeclaration
The warning is:
stime.c:85: warning: `tzname' redeclared without dllimport attribute: previous dllimport ignored

The second half of this section of #ifdefery implies that the MinGW
build isn't going to reference a function called `tzname' anyway
(because it's going to use `_tzname' instead).  So why should a MinGW
build want any declaration of `tzname' at all?

As far as a declaration of `_tzname' is concerned, there are no
more #includes after this, so we must have already got a declaration
of `_tzname' from one of the preceding includes.

* libguile/stime.c (tzname): Remove declaration in __MINGW32__ build.
  It isn't needed and causes a build warning.